refactor: extract AddOrderPage from App
