+++ src/components/pages/add-order-page/index.js
@@ -1,0 +1,3 @@
+import ConnectedAddOrderPage from './AddOrderPage'
+
+export default ConnectedAddOrderPage
